refactor: CarouselItems 코드 분기
