+++ src/App.vue
@@ -28,7 +28,7 @@ const TEST = [
     <h1 class="text-3xl font-bold mb-4">Hello Minimal Carousel</h1>
 
     <div class="w-2/4 h-full">
-      <MinimalCarousel :imageItems="TEST" pagination dynamicBullets />
+      <MinimalCarousel :imageItems="TEST" />
     </div>
   </div>
 </template>
