+++ src/components/MinimalCarousel.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
-import { computed, onMounted, ref, toRefs, watch } from 'vue';
-import type { ImageItemsType } from '@/types/items.type';
+import { computed, ref, toRefs, watch } from 'vue';
 import CarouselController from './CarouselController.vue';
 import CarouselPagination from './CarouselPagination.vue';
 import CarouselScrollbar from './CarouselScrollbar.vue';
-import CarouselContents from './CarouselContents.vue';
+import CarouselItems from './CarouselItems.vue';
+import type { ImageItemsType } from '@/types/items.type';
 
 const props = withDefaults(
   defineProps<{
@@ -21,7 +21,6 @@ const props = withDefaults(
     stopOnLastSlide?: boolean;
     effectFade?: boolean;
     keyboardControl?: boolean;
-    parallax?: boolean;
     contents?: {
       title: string;
       subTitle: string;
@@ -35,27 +34,21 @@ const props = withDefaults(
     dynamicBullets: false,
     scrollbar: false,
     autoPlay: false,
-    autoPlayDuration: 1000,
+    autoPlayDuration: 2500,
     disableOnInteraction: true,
     pauseOnMouseEnter: false,
     stopOnLastSlide: false,
     effectFade: false,
     keyboardControl: false,
-    parallax: false,
     contents: undefined,
   },
 );
-
 const { imageItems, scrollbar, autoPlay, pagination } = toRefs(props);
 
 const showScrollbar = computed(() => scrollbar.value ?? false);
-const carouselRef = ref<HTMLElement | null>(null);
 const currentIndex = ref(0);
-const isDragging = ref(false);
-const startX = ref(0);
-const startTranslateX = ref(0);
-
 const buttonClicked = ref(false);
+
 function handleButtonClick() {
   buttonClicked.value = true;
 }
@@ -91,57 +84,7 @@ function prevHandler() {
   }
 }
 
-// 드래그 시작
-function dragStart(event: MouseEvent) {
-  if (carouselRef.value) {
-    isDragging.value = true;
-    startX.value = event.pageX;
-    startTranslateX.value = currentIndex.value * -100;
-  }
-}
-
-// 드래그 중
-function dragging(event: MouseEvent) {
-  if (!isDragging.value || !carouselRef.value) return;
-
-  const dragX = event.pageX;
-  const walk = ((dragX - startX.value) / window.innerWidth) * 100; // 드래그 거리 계산
-
-  carouselRef.value.style.transition = 'none'; // 드래그 중에는 애니메이션 비활성화
-  carouselRef.value.style.transform = `translateX(${
-    startTranslateX.value + walk
-  }%)`; // 이동
-}
-
 const autoPlayInterval = ref<number | null>(null);
-
-// 드래그 종료
-function dragStop(event: MouseEvent) {
-  if (!isDragging.value) return;
-
-  const dragStandardDistance = 4; // 드래그 기준 거리, 현재 %로 계산됨
-  const dragX = event.pageX;
-  const walk = ((dragX - startX.value) / window.innerWidth) * 100; // 드래그 거리 계산
-
-  // 드래그 거리 기준으로 이전/다음 이미지로 이동
-  if (
-    walk < -dragStandardDistance &&
-    currentIndex.value < imageItems.value.length - 1
-  ) {
-    nextHandler(); // 다음 이미지로 이동
-  } else if (walk > dragStandardDistance && currentIndex.value > 0) {
-    prevHandler(); // 이전 이미지로 이동
-  }
-
-  if (carouselRef.value) {
-    carouselRef.value.style.transition = 'transform 0.4s ease'; // 이미지 드래그할때 여러 애니메이션을 줄 수 있음 일반적으로 ease
-    // 드래그 종료 시 최종 위치를 정해줘야 그 이미지를 볼 수 있기 떄문에
-    carouselRef.value.style.transform = `translateX(${
-      currentIndex.value * -100
-    }%)`; // 최종 위치 설정
-    isDragging.value = false;
-  }
-}
 
 function goToImage(index: number) {
   currentIndex.value = index;
@@ -194,72 +137,23 @@ watch(
   },
   { immediate: true },
 );
-
-function effectFadeStyle(index: number) {
-  return `${
-    currentIndex.value === index
-      ? 'opacity-100 transition-opacity duration-1000 ease-in-out'
-      : 'opacity-0'
-  }`;
-}
-
-function handleKeyDown(event: KeyboardEvent) {
-  if (props.keyboardControl) {
-    if (event.key === 'ArrowRight') {
-      nextHandler();
-    } else if (event.key === 'ArrowLeft') {
-      prevHandler();
-    }
-  }
-}
-
-onMounted(() => {
-  if (carouselRef.value) {
-    carouselRef.value.focus(); // 자동 포커스
-  }
-});
 </script>
 
 <template>
   <div class="overflow-hidden relative flex-shrink-0">
-    <ul
-      ref="carouselRef"
-      class="flex transition-transform duration-400"
-      :style="{ transform: `translateX(${currentIndex * -100}%)` }"
-      @mousedown="dragStart"
-      @mouseup="dragStop"
-      @mouseleave="dragStop"
-      @mousemove="dragging"
-      @keydown="handleKeyDown"
-      tabindex="0"
-    >
-      <li
-        v-for="(image, index) in imageItems"
-        :key="index"
-        :class="`flex-shrink-0 w-full h-full ${
-          effectFade && effectFadeStyle(index)
-        }`"
-        @click="handleInteraction"
-        @mouseenter="handleMouseEnter"
-        @mouseleave="handleMouseEnter"
-      >
-        <CarouselContents
-          v-if="props.contents && props.contents.length > index"
-          :contents="props.contents[index]"
-          :index="index"
-          :current-index="currentIndex"
-          :image="image"
-        />
-        <div v-else>
-          <img
-            :src="image.link"
-            :alt="image.name"
-            draggable="false"
-            class="w-full h-full object-cover"
-          />
-        </div>
-      </li>
-    </ul>
+    <CarouselItems
+      :imageItems="imageItems"
+      :contents="props.contents"
+      :currentIndex="currentIndex"
+      :keyboardControl="props.keyboardControl"
+      :effectFade="props.effectFade"
+      :disableOnInteraction="props.disableOnInteraction"
+      :pauseOnMouseEnter="props.pauseOnMouseEnter"
+      :handleInteraction="handleInteraction"
+      :handleMouseEnter="handleMouseEnter"
+      @prev="prevHandler"
+      @next="nextHandler"
+    />
 
     <CarouselController
       :imageItemsLength="imageItems.length"
